Added HaloSign
First iteration of halosign.

I was able to have it sample the texture but when I have it try to sample the other texture as a sub UV it picks up the color but it doesn't give me the SubUVIndex. We're so close I can taste it. Let's try and figure it out on thursday.