Watch Additon
Added Widget folder, you'll see in the widget folder I have a font folder, and Icon folder. I added a font family to the project for the points on the watch. Icon folder should be self explanitory. Then I added the watch BP and widget watch.

In the BP_Controller it's calling a the widget, It was super large so I had to re-size it in the BP_Controller.